HashiCorpVault and Exception Handler

+++ src/main/java/com/onlineshop/service/ProductService.java
@@ -1,5 +1,7 @@
 package com.onlineshop.service;
 
+
+import com.onlineshop.exception.ProductNotFoundException;
 import com.onlineshop.model.product;
 import com.onlineshop.repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -9,17 +11,21 @@ import java.util.List;
 
 @Transactional
 @Service
-
 public class ProductService {
     private final ProductRepository productRepository;
 
+
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
+
     }
+
     public product createProduct(product product) {
-        return  productRepository.save(product) ;
+        return productRepository.save(product);
     }
-    public List<product> findAll(){
+
+    public List<product> findAll() {
+
         return productRepository.findAll();
     }
 
@@ -27,11 +33,12 @@ public class ProductService {
         if (productRepository.existsById(ID)) {
             productEntity.setId(ID);
             return productRepository.save(productEntity);
-        }
-        else throw new RuntimeException("ID Not found") ;
+        } else throw new ProductNotFoundException(ID);
     }
 
     public void deleteProduct(String ID) {
-        productRepository.deleteById(ID);
+        if (productRepository.existsById(ID)) {
+            productRepository.deleteById(ID);
+        } else throw new ProductNotFoundException(ID);
     }
 }
